CTECH-1344: Enables notifications for PR's and Issues on forked PR's (#314)
When forked PR's are created, notifications are meant to be raised. With the current setup, this is not possible. This PR changes the workflow actions from pull_request to pull_request_target; pull_request_target runs in the context of the target branch, giving it access to secrets. Because it is run on the target branch, the PR cannot exfiltrate any secrets or have changes to workflows run in the context of the target repository. Reviewers still need to be vigilent of changes from unknown sources, especially around workflows.

Signed-off-by: Paul Saunders <pms1969@gmail.com>